Refactor Plugin to Match Webmentions Refactor
