+++ languages/semantic_linkbacks.pot
@@ -1,138 +1,140 @@
-# Copyright (C) 2015 pfefferle
+# Copyright (C) 2016 pfefferle
 # This file is distributed under the same license as the Semantic-Linkbacks package.
 msgid ""
 msgstr ""
 "Project-Id-Version: Semantic-Linkbacks 3.2.0\n"
 "Report-Msgid-Bugs-To: "
-"https://wordpress.org/support/plugin/semantic-linkbacks\n"
-"POT-Creation-Date: 2015-09-14 20:23:55+00:00\n"
+"http://wordpress.org/support/plugin/semantic-linkbacks\n"
+"POT-Creation-Date: 2016-06-29 03:34:03+00:00\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"PO-Revision-Date: 2015-MO-DA HO:MI+ZONE\n"
+"PO-Revision-Date: 2016-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
-"X-Generator: grunt-wp-i18n 0.5.3\n"
+"X-Generator: grunt-wp-i18n 0.4.9\n"
 
-#: semantic-linkbacks.php:180
+#: includes/class-linkbacks-handler.php:169
 msgid "%1$s mentioned %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:182
+#: includes/class-linkbacks-handler.php:171
 msgid "%1$s replied to %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:183
+#: includes/class-linkbacks-handler.php:172
 msgid "%1$s reposted %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:184
+#: includes/class-linkbacks-handler.php:173
 msgid "%1$s liked %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:185
+#: includes/class-linkbacks-handler.php:174
 msgid "%1$s favorited %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:186
+#: includes/class-linkbacks-handler.php:175
 msgid "%1$s tagged %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:187
+#: includes/class-linkbacks-handler.php:176
 msgid "%1$s bookmarked %2$s on <a href=\"%3$s\">%4$s</a>"
 msgstr ""
 
-#: semantic-linkbacks.php:188
+#: includes/class-linkbacks-handler.php:177
 msgid "%1$s is <strong>attending</strong>"
 msgstr ""
 
-#: semantic-linkbacks.php:189
+#: includes/class-linkbacks-handler.php:178
 msgid "%1$s is <strong>not attending</strong>"
 msgstr ""
 
-#: semantic-linkbacks.php:190
+#: includes/class-linkbacks-handler.php:179
 msgid "Maybe %1$s will be <strong>attending</strong>"
 msgstr ""
 
-#: semantic-linkbacks.php:191
+#: includes/class-linkbacks-handler.php:180
 msgid "%1$s is <strong>invited</strong>"
 msgstr ""
 
-#: semantic-linkbacks.php:192
+#: includes/class-linkbacks-handler.php:181
 msgid "%1$s <strong>tracks</strong> this event"
 msgstr ""
 
-#: semantic-linkbacks.php:206
+#: includes/class-linkbacks-handler.php:195
 msgid "Mention"
 msgstr ""
 
-#: semantic-linkbacks.php:208
+#: includes/class-linkbacks-handler.php:197
 msgid "Reply"
 msgstr ""
 
-#: semantic-linkbacks.php:209
+#: includes/class-linkbacks-handler.php:198
 msgid "Repost"
 msgstr ""
 
-#: semantic-linkbacks.php:210
+#: includes/class-linkbacks-handler.php:199
 msgid "Like"
 msgstr ""
 
-#: semantic-linkbacks.php:211
+#: includes/class-linkbacks-handler.php:200
 msgid "Favorite"
 msgstr ""
 
-#: semantic-linkbacks.php:212
+#: includes/class-linkbacks-handler.php:201
 msgid "Tag"
 msgstr ""
 
-#: semantic-linkbacks.php:213
+#: includes/class-linkbacks-handler.php:202
 msgid "Bookmark"
 msgstr ""
 
-#: semantic-linkbacks.php:214 semantic-linkbacks.php:215
-#: semantic-linkbacks.php:216 semantic-linkbacks.php:217
-#: semantic-linkbacks.php:218
+#: includes/class-linkbacks-handler.php:203
+#: includes/class-linkbacks-handler.php:204
+#: includes/class-linkbacks-handler.php:205
+#: includes/class-linkbacks-handler.php:206
+#: includes/class-linkbacks-handler.php:207
 msgid "RSVP"
 msgstr ""
 
-#: semantic-linkbacks.php:232
+#: includes/class-linkbacks-handler.php:221
 msgid "this Article"
 msgstr ""
 
-#: semantic-linkbacks.php:234
+#: includes/class-linkbacks-handler.php:223
 msgid "this Aside"
 msgstr ""
 
-#: semantic-linkbacks.php:235
+#: includes/class-linkbacks-handler.php:224
 msgid "this Chat"
 msgstr ""
 
-#: semantic-linkbacks.php:236
+#: includes/class-linkbacks-handler.php:225
 msgid "this Gallery"
 msgstr ""
 
-#: semantic-linkbacks.php:237
+#: includes/class-linkbacks-handler.php:226
 msgid "this Link"
 msgstr ""
 
-#: semantic-linkbacks.php:238
+#: includes/class-linkbacks-handler.php:227
 msgid "this Image"
 msgstr ""
 
-#: semantic-linkbacks.php:239
+#: includes/class-linkbacks-handler.php:228
 msgid "this Quote"
 msgstr ""
 
-#: semantic-linkbacks.php:240
+#: includes/class-linkbacks-handler.php:229
 msgid "this Status"
 msgstr ""
 
-#: semantic-linkbacks.php:241
+#: includes/class-linkbacks-handler.php:230
 msgid "this Video"
 msgstr ""
 
-#: semantic-linkbacks.php:242
+#: includes/class-linkbacks-handler.php:231
 msgid "this Audio"
 msgstr ""
